% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{cookie_placeholders}
\alias{cookie_placeholders}
\alias{default_cookie_getter}
\alias{default_cookie_setter}
\title{Place Holder Database Cookie Handlers}
\usage{
default_cookie_getter()

default_cookie_setter(user, session)
}
\arguments{
\item{user}{character string representing the user id}

\item{session}{character string representing the session id}
}
\value{
`default_cookie_setter` returns a data.frame with zero rows and two columns: `user` and `session`
}
\description{
These functions silently reject all cookies during login and ignore requests to save session
ids to a database.  Use these if you don't want to allow automatic login with cookies.  These are
default values for `callModule(shinyauthr::login, ...)`.  Replace them with your own functions which
get and set the user and session ids in a database.
}
\keyword{internal}
