% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{js_cookie_to_r_code}
\alias{js_cookie_to_r_code}
\title{Generate required Javascript}
\usage{
js_cookie_to_r_code(id, expire_days = 7)
}
\arguments{
\item{id}{name of input object to hold the cookie value}

\item{expire_days}{number of days to ask browser to retain cookie}
}
\value{
Character string of Javascript code
}
\description{
This is used to generate the Javascript which gets and sets the cookies in the
user's browser.  We need to glue it together to handle the variable expiry time and
to get the correct module namespace.  This function is used internally only.
}
\examples{
\dontrun{
shinyjs::extendShinyjs(
  text = js_cookie_to_r_code(ns("jscookie")),
  functions = c("getcookie", "setcookie", "rmcookie")
)
}
}
\keyword{internal}
