% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{js_return_click}
\alias{js_return_click}
\title{Javascript code to trigger login button with return key}
\usage{
js_return_click(idpassword, idbutton)
}
\arguments{
\item{idpassword}{name of password input, with correct namespace}

\item{idbutton}{name of action button, with correct namespace}
}
\description{
Javascript code to trigger login button with return key
}
\keyword{internal}
