% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{jscookie_script}
\alias{jscookie_script}
\title{Javascript code for managing cookies}
\usage{
jscookie_script()
}
\description{
This code is minified from the GitHub project js-cookie/js-cookie.  We include it here
to make the shinyauthr package more self-contained
}
\keyword{internal}
