% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{login server module (deprecated)}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{data}{data frame or tibble containing usernames, passwords and other user data}

\item{user_col}{bare (unquoted) column name containing usernames}

\item{pwd_col}{bare (unquoted) column name containing passwords}

\item{sodium_hashed}{have the passwords been hash encrypted using the sodium package? defaults to FALSE}

\item{hashed}{Deprecated. shinyauthr now uses the sodium package for password hashing and decryption. If you have previously hashed your passwords with the digest package to use with shinyauthr please re-hash them with sodium for decryption to work.}

\item{algo}{Deprecated}

\item{log_out}{[reactive] supply the returned reactive from \link{logout} here to trigger a user logout}

\item{sessionid_col}{bare (unquoted) column name containing session ids}

\item{cookie_getter}{a function that returns a data.frame with at least two columns: user and session}

\item{cookie_setter}{a function with two parameters: user and session.  The function must save these to a database.}

\item{reload_on_logout}{should app force reload on logout?}
}
\value{
The module will return a reactive 2 element list to your main application.
  First element \code{user_auth} is a boolean inditcating whether there has been
  a successful login or not. Second element \code{info} will be the data frame provided
  to the function, filtered to the row matching the successfully logged in username.
  When \code{user_auth} is FALSE \code{info} is NULL.
}
\description{
Shiny authentication module for use with \link{loginUI}
}
\details{
Call via \code{shiny::callModule(shinyauthr::login, "id", ...)}

This function is now deprecated in favour of \link{loginServer} which uses shiny's new \link[shiny]{moduleServer} 
method as opposed to the \link[shiny]{callModule} method used by this function. 
See the \link{loginServer} documentation For details on how to migrate.
}
\examples{
\dontrun{
user_credentials <- shiny::callModule(
  login,
  id = "login",
  data = user_base,
  user_col = user,
  pwd_col = password,
  log_out = reactive(logout_init())
)
}

}
