% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{loginUI}
\alias{loginUI}
\title{login UI module}
\usage{
loginUI(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Shiny UI
}
\description{
Shiny UI Module for use with \link{login}
}
\details{
Call via \code{loginUI("your_id")}
}
\author{
Paul Campbell, \email{pacampbell91@gmail.com}
}
