% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logout}
\alias{logout}
\title{logout server module}
\usage{
logout(input, output, session, active)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{active}{[reactive] supply the returned \code{user_auth} boolean reactive from \link{login}
here to hide/show the logout button}
}
\value{
The reactive output of this module should be supplied as the \code{log_out} argument to the
  \link{login} module to trigger the logout process
}
\description{
Shiny authentication module for use with \link{logoutUI}
}
\details{
Call via \code{shiny::callModule(logout, "your_id", ...)}
}
\examples{
\dontrun{
  logout_init <- shiny::callModule(logout, "logout",
                                   active = reactive(user_credentials()$user_auth))
}

}
