% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logoutServer}
\alias{logoutServer}
\title{logout server module}
\usage{
logoutServer(id, active, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function}

\item{active}{\code{reactive} supply the returned \code{user_auth} boolean reactive from \link{loginServer}
here to hide/show the logout button}

\item{...}{arguments passed to \link[shinyjs]{toggle}}
}
\value{
The reactive output of this module should be supplied as the \code{log_out} argument to the
  \link{loginServer} module to trigger the logout process
}
\description{
Shiny authentication module for use with \link{logoutUI}
}
\details{
Call via \code{logoutServer(id = "id", ...)}
}
\examples{
\dontrun{
logout_init <- logoutServer(
  id = "logout",
  active = reactive(credentials()$user_auth)
)
}

}
