% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{randomString}
\alias{randomString}
\title{Generate random string}
\usage{
randomString(n = 64)
}
\arguments{
\item{n}{length of string}
}
\value{
A random character string.
}
\description{
This function is used to generate random session ids.
}
\keyword{internal}
