% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runNavbarPageExample}
\alias{runNavbarPageExample}
\title{Run navbarPage example}
\usage{
runNavbarPageExample()
}
\description{
Launch example shiny navbarPage app using shinyauthr authentication modules
}
