% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runShinyDashboardExample}
\alias{runShinyDashboardExample}
\title{Run shinyauthr example}
\usage{
runShinyDashboardExample()
}
\description{
Launch example shinydashboard using shinyauthr authentication modules
}
