% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runShinyExample}
\alias{runShinyExample}
\title{Run shinyauthr example}
\usage{
runShinyExample()
}
\description{
Launch example shiny dashboard using shinyauthr authentication modules
}
\author{
Paul Campbell, \email{pacampbell91@gmail.com}
}
