% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logoutServer}
\alias{logoutServer}
\title{logout server module}
\usage{
logoutServer(id, active, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function}

\item{active}{\code{reactive} supply the returned \code{user_auth} boolean reactive from \link{loginServer}
here to hide/show the logout button}

\item{...}{arguments passed to \link[shinyjs]{toggle}}
}
\value{
Reactive boolean, to be supplied as the \code{log_out} argument of the
  \link{loginServer} module to trigger the logout process
}
\description{
Shiny authentication module for use with \link{logoutUI}
}
\details{
This module uses shiny's new \link[shiny]{moduleServer} method as opposed to the \link[shiny]{callModule}
method used by the now deprecated \link{login} function and must be called differently in your app.
For details on how to migrate see the 'Migrating from callModule to moduleServer' section of 
\href{https://shiny.rstudio.com/articles/modules.html}{Modularizing Shiny app code}.
}
\examples{
\dontrun{
logout_init <- logoutServer(
  id = "logout",
  active = reactive(credentials()$user_auth)
)
}

}
