% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logoutUI}
\alias{logoutUI}
\title{logout UI module}
\usage{
logoutUI(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Shiny UI action button
}
\description{
Shiny UI Module for use with \link{logout}
}
\details{
Call via \code{logoutUI("your_id")}
}
\author{
Paul Campbell, \email{pacampbell91@gmail.com}
}
