% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logoutUI}
\alias{logoutUI}
\title{logout UI module}
\usage{
logoutUI(
  id,
  label = "Log out",
  icon = NULL,
  class = "btn-danger",
  style = "color: white;"
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's server function}

\item{label}{label for the logout button}

\item{icon}{An optional \code{\link[shiny]{icon}} to appear on the button.}

\item{class}{bootstrap class for the logout button}

\item{style}{css styling for the logout button}
}
\value{
Shiny UI action button
}
\description{
Shiny UI Module for use with \link{logout}
}
